Fix back button in article breadcrumbs
